+++ src/contentful/cache.ts
@@ -4,6 +4,9 @@ import { Entry } from "contentful";
 export interface PageCache {
     getSlugOfPage: (id: string) => string;
     getTitleOfPage: (id: string) => string;
+    getTitleOfPageBySlug: (slug: string) => string | null;
+    getIdOfPageBySlug: (slug: string) => string | null;
+    isSlugExisting: (slug: string) => boolean;
 }
 
 export const getPageCache = (pages: Entry<PageForCache>[]): PageCache => {
@@ -28,8 +31,47 @@ export const getPageCache = (pages: Entry<PageForCache>[]): PageCache => {
         return titleCache[id];
     };
 
+    const getTitleOfPageBySlug = (slug: string): string | null => {
+        for (let id in slugCache) {
+            const currentSlug = slugCache[id];
+
+            if (currentSlug === slug) {
+                return getTitleOfPage(id);
+            }
+        }
+
+        return null;
+    };
+
+    const getIdOfPageBySlug = (slug: string): string | null => {
+        for (let id in slugCache) {
+            const currentSlug = slugCache[id];
+
+            if (currentSlug === slug) {
+                return id;
+            }
+        }
+
+        return null;
+    };
+
+    const isSlugExisting = (slug: string): boolean => {
+        for (let id in slugCache) {
+            const currentSlug = slugCache[id];
+
+            if (currentSlug === slug) {
+                return true;
+            }
+        }
+
+        return false;
+    };
+
     return {
         getSlugOfPage,
         getTitleOfPage,
+        getTitleOfPageBySlug,
+        getIdOfPageBySlug,
+        isSlugExisting,
     };
 };